refactor: Add reusable keywords for opening URLs in different browsers
This commit adds a new file `keywords.robot` which contains reusable keywords for opening URLs in different browsers. The keywords `Open Url With Chrome` and `Open Url With Firefox` are added to open URLs in Chrome and Firefox respectively. The `Open Url` keyword is modified to accept the browser as an argument and open the URL accordingly. Additionally, the `Enter Keyword` and `Click Search Button` keywords are added to interact with the search functionality on the page.

These changes improve code reusability and make it easier to open URLs in different browsers without duplicating code.